// app/server/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const MemoryType = {
  NOTE: 'NOTE',
  LINK: 'LINK',
  IMAGE: 'IMAGE',
  DOC: 'DOC'
} as const

export type MemoryType = (typeof MemoryType)[keyof typeof MemoryType]


export const ShareVisibility = {
  PUBLIC: 'PUBLIC',
  SECRET: 'SECRET',
  RESTRICTED: 'RESTRICTED'
} as const

export type ShareVisibility = (typeof ShareVisibility)[keyof typeof ShareVisibility]